image_generating_with_scraped_quotes/pyproject.toml
[project]
name = "image-generating-with-scraped-quotes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "lxml>=6.0.0",
    "pillow>=11.3.0",
    "requests>=2.32.4",
]
